requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai
langgraph>=0.4.0
sentence-transformers
faiss-cpu
fastapi
uvicorn